overhaul outputs and share for v4 + v3
