+++ packages/commutable/src/types.ts
@@ -1,9 +1,6 @@
-import {
-  Map as ImmutableMap,
-  List as ImmutableList,
-  Record,
-  RecordOf
-} from "immutable";
+import { Map as ImmutableMap, List as ImmutableList, Record } from "immutable";
+
+import { ImmutableOutput } from "./outputs";
 
 // Mutable JSON types
 export type PrimitiveImmutable = string | number | boolean | null;
@@ -20,8 +17,6 @@ export type ImmutableJSONType =
   | ImmutableJSONList;
 interface ImmutableJSONMap extends ImmutableMap<string, ImmutableJSONType> {}
 interface ImmutableJSONList extends ImmutableList<ImmutableJSONType> {}
-
-export type ExecutionCount = number | null;
 
 export type MimeBundle = JSONObject;
 
@@ -51,10 +46,7 @@ export type ImmutableCodeCell = ImmutableMap<string, any>;
 export type ImmutableMarkdownCell = ImmutableMap<string, any>;
 export type ImmutableRawCell = ImmutableMap<string, any>;
 export type ImmutableCell = ImmutableCodeCell | ImmutableMarkdownCell;
-export type ImmutableOutput = ImmutableMap<string, any>;
 export type ImmutableOutputs = ImmutableList<ImmutableOutput>;
-
-export type ImmutableMimeBundle = ImmutableMap<string, any>;
 
 export type ImmutableCellOrder = ImmutableList<CellID>;
 export type ImmutableCellMap = ImmutableMap<CellID, ImmutableCell>;
