+++ packages/commutable/src/v4.ts
@@ -18,9 +18,7 @@ import {
   Map as ImmutableMap,
   fromJS as immutableFromJS,
   List as ImmutableList,
-  Set as ImmutableSet,
-  Record,
-  RecordOf
+  Set as ImmutableSet
 } from "immutable";
 
 import {
@@ -31,25 +29,27 @@ import {
   ImmutableMarkdownCell,
   ImmutableRawCell,
   ImmutableCell,
-  ImmutableMimeBundle,
-  ExecutionCount,
   JSONObject,
   JSONType,
   MultiLineString
 } from "./types";
 
-import { appendCell, CellStructure } from "./structures";
+import {
+  createImmutableMimeBundle,
+  ImmutableOutput,
+  makeExecuteResult,
+  makeDisplayData,
+  makeStreamOutput,
+  makeErrorOutput,
+  demultiline,
+  remultiline,
+  isJSONKey,
+  MimeBundle,
+  ImmutableMimeBundle,
+  ExecutionCount
+} from "./outputs";
 
-//
-// MimeBundle example (disk format)
-//
-// {
-//   "application/json": {"a": 3, "b": 2},
-//   "text/html": ["<p>\n", "Hey\n", "</p>"],
-//   "text/plain": "Hey"
-// }
-//
-export type MimeBundle = { [key: string]: string | string[] | Object };
+import { appendCell, CellStructure } from "./structures";
 
 /** * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
  *                             Output Types
@@ -115,146 +115,6 @@ export type Notebook = {
   nbformat: 4;
   nbformat_minor: number;
 };
-
-export const demultiline = (s: string | string[]) =>
-  Array.isArray(s) ? s.join("") : s;
-
-/**
- * Split string into a list of strings delimited by newlines
- */
-export const remultiline = (s: string | string[]): string[] =>
-  Array.isArray(s) ? s : s.split(/(.+?(?:\r\n|\n))/g).filter(x => x !== "");
-
-export const isJSONKey = (key: string) =>
-  /^application\/(.*\+)?json$/.test(key);
-
-export const cleanMimeData = (
-  key: string,
-  data: string | string[] | object
-) => {
-  // See https://github.com/jupyter/nbformat/blob/62d6eb8803616d198eaa2024604d1fe923f2a7b3/nbformat/v4/nbformat.v4.schema.json#L368
-  if (isJSONKey(key)) {
-    // Data stays as is for JSON types
-    return data;
-  }
-
-  if (typeof data === "string" || Array.isArray(data)) {
-    return demultiline(data);
-  }
-
-  throw new TypeError(
-    `Data for ${key} is expected to be a string or an Array of strings`
-  );
-};
-
-// Map over all the mimetypes, turning them into our in-memory format
-//
-// {
-//   "application/json": {"a": 3, "b": 2},
-//   "text/html": ["<p>\n", "Hey\n", "</p>"],
-//   "text/plain": "Hey"
-// }
-//
-// to
-//
-// {
-//   "application/json": {"a": 3, "b": 2},
-//   "text/html": "<p>\nHey\n</p>",
-//   "text/plain": "Hey"
-// }
-//
-export const cleanMimeAtKey = (
-  mimeBundle: MimeBundle,
-  previous: ImmutableMimeBundle,
-  key: string
-): ImmutableMimeBundle =>
-  previous.set(key, cleanMimeData(key, mimeBundle[key]));
-
-export const createImmutableMimeBundle = (
-  mimeBundle: MimeBundle
-): ImmutableMimeBundle =>
-  Object.keys(mimeBundle).reduce(
-    cleanMimeAtKey.bind(null, mimeBundle),
-    ImmutableMap()
-  );
-
-export const sanitize = (o: ExecuteResult | DisplayData) =>
-  o.metadata ? { metadata: immutableFromJS(o.metadata) } : {};
-
-/** ExecuteResult Record Boilerplate */
-type ExecuteResultParams = {
-  output_type: "execute_result";
-  execution_count: ExecutionCount;
-  data: ImmutableMimeBundle;
-  metadata?: any;
-};
-
-export const makeExecuteResult = Record<ExecuteResultParams>({
-  output_type: "execute_result",
-  execution_count: null,
-  data: ImmutableMap(),
-  metadata: ImmutableMap()
-});
-
-type ImmutableExecuteResult = RecordOf<ExecuteResultParams>;
-
-/** DisplayData Record Boilerplate */
-
-type DisplayDataParams = {
-  output_type: "display_data";
-  data: ImmutableMimeBundle;
-  metadata?: any;
-};
-
-export const makeDisplayData = Record<DisplayDataParams>({
-  output_type: "display_data",
-  data: ImmutableMap(),
-  metadata: ImmutableMap()
-});
-
-type ImmutableDisplayData = RecordOf<DisplayDataParams>;
-
-/** StreamOutput Record Boilerplate */
-
-type StreamOutputParams = {
-  output_type: "stream";
-  name: "stdout" | "stderr";
-  text: string;
-};
-
-export const makeStreamOutput = Record<StreamOutputParams>({
-  output_type: "stream",
-  name: "stdout",
-  text: ""
-});
-
-type ImmutableStreamOutput = RecordOf<StreamOutputParams>;
-
-/** ErrorOutput Record Boilerplate */
-
-type ErrorOutputParams = {
-  output_type: "error";
-  ename: string;
-  evalue: string;
-  traceback: ImmutableList<string>;
-};
-
-export const makeErrorOutput = Record<ErrorOutputParams>({
-  output_type: "error",
-  ename: "",
-  evalue: "",
-  traceback: ImmutableList()
-});
-
-type ImmutableErrorOutput = RecordOf<ErrorOutputParams>;
-
-//////////////
-
-type ImmutableOutput =
-  | ImmutableExecuteResult
-  | ImmutableDisplayData
-  | ImmutableStreamOutput
-  | ImmutableErrorOutput;
 
 export const createImmutableOutput = (output: Output): ImmutableOutput => {
   switch (output.output_type) {
